Support for ActiveRecord 8.0
