+++ Appraisals
@@ -1,3 +1,8 @@
+appraise "ar-8.0" do
+  gem "activerecord", "~> 8.0.0"
+  gem "sqlite3", "~> 2.4.1"
+end
+
 appraise "ar-7.2" do
   gem "activerecord", "~> 7.2.1"
   gem "sqlite3", "~> 1.7.3"
