+++ gemfiles/ar_7.1.gemfile.lock
@@ -1,8 +1,8 @@
 PATH
   remote: ..
   specs:
-    otr-activerecord (2.3.0)
-      activerecord (>= 6.0, < 7.3)
+    otr-activerecord (2.4.0)
+      activerecord (>= 6.0, < 8.1)
       hashie-forbidden_attributes (~> 0.1)
 
 GEM
