+++ gemfiles/ar_8.0.gemfile
@@ -1,0 +1,15 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "rake"
+gem "appraisal", "~> 2.5"
+gem "rack", "~> 3.1"
+gem "activerecord", "~> 8.0.0"
+gem "sqlite3", "~> 2.4.1"
+
+group :test do
+  gem "minitest"
+end
+
+gemspec path: "../"
