+++ gemfiles/ar_8.0.gemfile.lock
@@ -8,14 +8,15 @@ PATH
 GEM
   remote: https://rubygems.org/
   specs:
-    activemodel (7.2.1)
-      activesupport (= 7.2.1)
-    activerecord (7.2.1)
-      activemodel (= 7.2.1)
-      activesupport (= 7.2.1)
+    activemodel (8.0.0.1)
+      activesupport (= 8.0.0.1)
+    activerecord (8.0.0.1)
+      activemodel (= 8.0.0.1)
+      activesupport (= 8.0.0.1)
       timeout (>= 0.4.0)
-    activesupport (7.2.1)
+    activesupport (8.0.0.1)
       base64
+      benchmark (>= 0.3)
       bigdecimal
       concurrent-ruby (~> 1.0, >= 1.3.1)
       connection_pool (>= 2.2.5)
@@ -25,11 +26,13 @@ GEM
       minitest (>= 5.1)
       securerandom (>= 0.3)
       tzinfo (~> 2.0, >= 2.0.5)
+      uri (>= 0.13.1)
     appraisal (2.5.0)
       bundler
       rake
       thor (>= 0.14.0)
     base64 (0.2.0)
+    benchmark (0.4.0)
     bigdecimal (3.1.8)
     concurrent-ruby (1.3.4)
     connection_pool (2.4.1)
@@ -37,32 +40,33 @@ GEM
     hashie (5.0.0)
     hashie-forbidden_attributes (0.1.1)
       hashie (>= 3.0)
-    i18n (1.14.5)
+    i18n (1.14.6)
       concurrent-ruby (~> 1.0)
-    logger (1.6.0)
-    minitest (5.24.0)
-    rack (3.1.4)
+    logger (1.6.3)
+    minitest (5.25.4)
+    rack (3.1.8)
     rake (13.2.1)
-    securerandom (0.3.1)
-    sqlite3 (1.7.3-aarch64-linux)
-    sqlite3 (1.7.3-arm64-darwin)
-    thor (1.3.1)
-    timeout (0.4.1)
+    securerandom (0.4.0)
+    sqlite3 (2.4.1-aarch64-linux-gnu)
+    sqlite3 (2.4.1-arm64-darwin)
+    thor (1.3.2)
+    timeout (0.4.2)
     tzinfo (2.0.6)
       concurrent-ruby (~> 1.0)
+    uri (1.0.2)
 
 PLATFORMS
   aarch64-linux
   arm64-darwin
 
 DEPENDENCIES
-  activerecord (~> 7.2.1)
+  activerecord (~> 8.0.0)
   appraisal (~> 2.5)
   minitest
   otr-activerecord!
   rack (~> 3.1)
   rake
-  sqlite3 (~> 1.7.3)
+  sqlite3 (~> 2.4.1)
 
 BUNDLED WITH
    2.5.11
