+++ lib/otr-activerecord/shim/v8.rb
@@ -2,7 +2,7 @@ require_relative 'base'
 
 module OTR
   module ActiveRecord
-    # Compatibility layer for ActiveRecord 7
+    # Compatibility layer for ActiveRecord 8
     class Shim < ShimBase
     end
   end
