+++ test/matrix
@@ -2,11 +2,13 @@
 # Read by bin/testall to test various combinations of Ruby, ActiveRecord, and datbases
 #
 
+ruby-3.3  ar-8.0  sqlite3
 ruby-3.3  ar-7.2  sqlite3
 ruby-3.3  ar-7.1  sqlite3
 ruby-3.3  ar-7.0  sqlite3
 ruby-3.3  ar-6.1  sqlite3
 
+ruby-3.2  ar-8.0  sqlite3
 ruby-3.2  ar-7.2  sqlite3
 ruby-3.2  ar-7.1  sqlite3
 ruby-3.2  ar-7.0  sqlite3
